restructured constants file
